refactor(capabilities): migrate the engine proxy and server caps to the struct-hosted actor macro
Move the EngineProxy and EngineServer #[actor] cardinality onto the cap
struct and relocate each impl NativeActor block (retagged #[runtime])
into its runtime.rs, per ADR-0123 (issue 2384). No behavior change.
