+++ crates/aether-capabilities/src/engine/proxy/mod.rs
@@ -79,271 +79,23 @@ pub use config::{EngineProxyConfig, HeartbeatParams};
 /// heartbeat handle) lives behind the one `feature = "runtime"` gate, so a
 /// transport-only build never names `EngineProxyState` nor pulls
 /// `aether_substrate` through this cap.
+#[actor(instanced)]
 pub struct EngineProxy;
 
 // The `#[actor]` / `#[handler]` attribute path stays always-on (the macro
 // divides what it emits). Everything that names an `aether_substrate` type —
 // the handler/init ctx, the runtime state, the connect/heartbeat helpers,
 // `Drop` — lives in the `runtime` module below, gated once by
-// `feature = "runtime"`; the `#[actor] impl` reaches all of it through the
-// single `use runtime::*` glob. The handler-signature kinds
-// (`ForwardEnvelope` / `RpcInboundReady` / …) stay always-on at file root —
-// the always-on `HandlesKind<K>` markers name them.
+// `feature = "runtime"`; the struct-hosted `#[actor]` reads that module's
+// `impl NativeActor` off disk to emit the identity. The handler-signature
+// kinds (`ForwardEnvelope` / `RpcInboundReady` / …) stay always-on at file
+// root — the always-on `HandlesKind<K>` markers name them.
 use aether_actor::actor;
 
-// The `runtime` module is this cap's private runtime-half namespace; the impl
-// reaches all of it (state, ctx types, connect/heartbeat helpers) through this
-// single seam, so the glob is intentional rather than a dozen one-line
-// imports.
-#[cfg(feature = "runtime")]
-#[allow(clippy::wildcard_imports)]
-use runtime::*;
-
-#[actor(instanced)]
-impl NativeActor for EngineProxy {
-    /// The runtime state this identity boots into (ADR-0122 split): the
-    /// per-engine outbound RPC connection plus the in-flight
-    /// reply-correlation table.
-    type State = EngineProxyState;
-    type Config = EngineProxyConfig;
-    const NAMESPACE: &'static str = "aether.engine.proxy";
-
-    fn init(
-        mut config: EngineProxyConfig,
-        ctx: &mut NativeInitCtx<'_>,
-    ) -> Result<EngineProxyState, BootError> {
-        let self_mailbox = ctx.self_id();
-        let mailer = ctx.mailer();
-        let wake_kind = KindId(<RpcInboundReady as Kind>::ID.0);
-
-        // A freshly-forked substrate (`spawned.is_some()`) may not
-        // have bound its RPC port yet, so the startup dial retries
-        // briefly. An adopted / externally-running substrate
-        // (`spawned.is_none()`) is dialed once — a refused
-        // connection there is a real error, not a startup race.
-        let retry = config.spawned.is_some();
-        let conn = match connect_proxy(
-            &config.rpc_addr,
-            &mailer,
-            self_mailbox,
-            wake_kind,
-            retry,
-            config.connect_budget,
-        ) {
-            Ok(conn) => conn,
-            Err(e) => {
-                // The proxy owns the child it was handed — a
-                // failed boot must not orphan the substrate.
-                if let Some(mut child) = config.spawned.take() {
-                    let _ = child.kill();
-                    let _ = child.wait();
-                }
-                return Err(BootError::Other(Box::new(e)));
-            }
-        };
-
-        tracing::info!(
-            target: "aether_substrate::engine_proxy",
-            engine_id = ?config.engine_id,
-            addr = %config.rpc_addr,
-            spawned = config.spawned.is_some(),
-            "engine proxy connected",
-        );
-
-        // Arm the liveness heartbeat, if configured. The sidecar
-        // thread fires an `EngineHeartbeatTick` at this proxy's own
-        // mailbox every `interval`; `on_heartbeat_tick` does the
-        // ping + miss accounting on the dispatcher thread (so the
-        // RPC write and all proxy state stay single-threaded).
-        let (heartbeat, miss_limit) = match config.heartbeat {
-            Some(params) if !params.interval.is_zero() && params.miss_limit > 0 => {
-                let handle = spawn_heartbeat(Arc::clone(&mailer), self_mailbox, params.interval);
-                (Some(handle), params.miss_limit)
-            }
-            _ => (None, 0),
-        };
-
-        Ok(EngineProxyState {
-            engine_id: config.engine_id,
-            mailer,
-            conn,
-            in_flight: HashMap::new(),
-            spawned: config.spawned,
-            missed_heartbeats: 0,
-            miss_limit,
-            heartbeat_seq: 0,
-            _heartbeat: heartbeat,
-        })
-    }
-
-    /// Relay one mail to the substrate as an RPC `Call`.
-    ///
-    /// # Agent
-    /// Hand the proxy a `ForwardEnvelope { mailbox, kind, payload }`
-    /// — the `mailbox` is the *substrate-local* recipient, `kind` +
-    /// `payload` the mail to deliver there. Any reply routes back to
-    /// the sender of this `ForwardEnvelope`.
-    #[handler]
-    fn on_forward(state: &mut Self::State, ctx: &mut NativeCtx<'_>, mail: ForwardEnvelope) {
-        let envelope = MailEnvelope {
-            to: MailboxAddress::local(mail.mailbox),
-            from: None,
-            kind: mail.kind,
-            correlation_id: None,
-            payload: mail.payload,
-        };
-        match state.conn.client.call(envelope) {
-            Ok(cid) => {
-                state.in_flight.insert(cid, ctx.reply_target());
-            }
-            Err(e) => {
-                tracing::warn!(
-                    target: "aether_substrate::engine_proxy",
-                    engine_id = ?state.engine_id,
-                    error = %e,
-                    "engine proxy: Call write failed; dropping forward",
-                );
-            }
-        }
-    }
-
-    /// Reader-sidecar wake. Drain every inbound frame.
-    ///
-    /// # Agent
-    /// Internal wake mail — not part of the proxy's external
-    /// surface. The reader thread fires this after pushing a frame;
-    /// the handler drains `conn.inbound` and routes each frame.
-    #[handler]
-    fn on_inbound_ready(state: &mut Self::State, ctx: &mut NativeCtx<'_>, _mail: RpcInboundReady) {
-        while let Ok(frame) = state.conn.inbound.try_recv() {
-            match frame {
-                WireFrame::ReplyEvent { cid, envelope } => state.route_reply(cid, envelope),
-                WireFrame::ReplyEnd { cid, result } => state.route_settled(cid, result),
-                // A `Pong` answers this proxy's heartbeat `Ping`
-                // (issue 1339): the substrate is alive. Clear the
-                // miss counter and report the liveness up to the
-                // engines cap so `list_engines` can show a fresh
-                // heartbeat age. The nonce is for log correlation
-                // only — any `Pong` is a liveness signal.
-                WireFrame::Pong(_nonce) => {
-                    state.missed_heartbeats = 0;
-                    state.report_alive(ctx);
-                }
-                WireFrame::Bye { reason } => {
-                    tracing::info!(
-                        target: "aether_substrate::engine_proxy",
-                        engine_id = ?state.engine_id,
-                        reason = %reason,
-                        "engine proxy: substrate closed the connection; shutting down",
-                    );
-                    // Tell the engines cap the engine is gone so it
-                    // drops the registry entry — without this the
-                    // proxy dies but `list_engines` keeps reporting
-                    // a corpse (issue 1339). The substrate closed the
-                    // connection on its own — a crash, not a
-                    // deliberate terminate; carry the `Bye` reason so
-                    // `list_engines` can show why.
-                    state.report_died(ctx, DeathReason::Crashed { detail: reason });
-                    ctx.shutdown();
-                    return;
-                }
-                // Hello / HelloAck / Call / Ping: a client-side proxy
-                // never expects these inbound. Drop with a debug
-                // line rather than warn-storming.
-                other => {
-                    tracing::debug!(
-                        target: "aether_substrate::engine_proxy",
-                        engine_id = ?state.engine_id,
-                        frame = ?other,
-                        "engine proxy: unexpected inbound frame; ignoring",
-                    );
-                }
-            }
-        }
-    }
-
-    /// Shut this proxy's substrate down.
-    ///
-    /// # Agent
-    /// Sent by the engines cap (`aether.engine`) on a terminate
-    /// request. The proxy self-shuts-down; its `Drop` SIGKILLs and
-    /// reaps the child substrate it forked (if any), and the
-    /// outbound RPC connection closes as the actor drops. The
-    /// `engine_id` field is ignored — a proxy only ever terminates
-    /// its own engine.
-    #[handler]
-    fn on_terminate(state: &mut Self::State, ctx: &mut NativeCtx<'_>, _mail: TerminateEngine) {
-        tracing::info!(
-            target: "aether_substrate::engine_proxy",
-            engine_id = ?state.engine_id,
-            "engine proxy: terminate requested; shutting down",
-        );
-        // No `report_died` here: the engines cap initiated this
-        // terminate and already dropped the registry entry, so the
-        // proxy reporting back would be a redundant (idempotent)
-        // no-op. The self-death paths (`Bye`, heartbeat timeout) are
-        // the ones the cap doesn't already know about.
-        ctx.shutdown();
-    }
-
-    /// Liveness-heartbeat timer wake (issue 1339).
-    ///
-    /// # Agent
-    /// Internal wake mail — not part of the proxy's external
-    /// surface. The heartbeat sidecar thread fires this every
-    /// interval. The handler counts the tick as an outstanding miss
-    /// (a `Pong` since the last tick would have cleared the
-    /// counter), and once `miss_limit` consecutive ticks go
-    /// unanswered it declares the engine dead: reports `EngineDied`
-    /// to the engines cap and self-shuts-down (its `Drop` SIGKILLs
-    /// the wedged child). Otherwise it sends a fresh `Ping`.
-    #[handler]
-    fn on_heartbeat_tick(
-        state: &mut Self::State,
-        ctx: &mut NativeCtx<'_>,
-        _mail: EngineHeartbeatTick,
-    ) {
-        state.heartbeat_seq += 1;
-        // A write failure means the socket is already broken — the
-        // reader sidecar will surface a `Bye` and `on_inbound_ready`
-        // handles the eviction. Count it as a miss and carry on so
-        // the miss-limit path also covers it (whichever fires first
-        // evicts; the cap side is idempotent).
-        if let Err(e) = state.conn.client.ping(state.heartbeat_seq) {
-            tracing::debug!(
-                target: "aether_substrate::engine_proxy",
-                engine_id = ?state.engine_id,
-                error = %e,
-                "engine proxy: heartbeat ping write failed",
-            );
-        }
-        state.missed_heartbeats += 1;
-        if state.missed_heartbeats >= state.miss_limit {
-            tracing::warn!(
-                target: "aether_substrate::engine_proxy",
-                engine_id = ?state.engine_id,
-                missed = state.missed_heartbeats,
-                miss_limit = state.miss_limit,
-                "engine proxy: heartbeat miss limit crossed; evicting engine",
-            );
-            state.report_died(
-                ctx,
-                DeathReason::Evicted {
-                    detail: format!(
-                        "heartbeat miss limit {} of {}",
-                        state.missed_heartbeats, state.miss_limit
-                    ),
-                },
-            );
-            ctx.shutdown();
-        }
-    }
-}
-
 // The runtime half — the whole `aether_substrate`-typed surface (imports,
-// `EngineProxyState`, its `Drop` + helper methods, `engine_cap_mailbox`) —
-// lives in `runtime.rs`, gated once here. The `#[actor] impl` above reaches it
-// through the `use runtime::*` glob.
+// `EngineProxyState`, its `Drop` + helper methods, `engine_cap_mailbox`) plus
+// the `#[runtime] impl NativeActor` — lives in `runtime.rs`, gated once here.
+// The struct-hosted `#[actor]` above reads it off disk to emit the identity.
 #[cfg(feature = "runtime")]
 mod runtime;
